audio: codec: separate common codec function and macro to another file
separate common codec function and macro to another file

Testing Done:
Bug Number:
Review URL:
Reviewed by:
